Updated to double intelligence from int level

+++ src/player/AbstractPlayerFactory.java
@@ -2,6 +2,6 @@ package player;
 
 public interface AbstractPlayerFactory {
 
-    public Player createPlayer(int intelligence);
+    public Player createPlayer(double intelligence);
 
 }
