+++ src/player/OpponentFactory.java
@@ -3,7 +3,7 @@ package player;
 public class OpponentFactory implements AbstractPlayerFactory{
 
     @Override
-    public Opponent createPlayer(int intelligence){
+    public Opponent createPlayer(double intelligence){
         Opponent player = new Opponent(intelligence);
         return player;
     }
